make testing more accessible with smaller files
